+++ declaration/out/declaration/index.d.ts
@@ -1,20 +1,20 @@
 import * as t from '../src';
-export declare const T5: t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const T5: t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
         }>;
     }>;
 }>;
-export declare const R1: t.RefinementT<t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const R1: t.RefinementT<t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
@@ -26,11 +26,11 @@ declare type R = {
     b: R | undefined | null;
 };
 export declare const Rec1: t.RecursiveType<t.Type<R, R, unknown>, R, R, unknown>;
-export declare const A1: t.ArrayT<t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const A1: t.ArrayT<t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
@@ -49,55 +49,55 @@ export declare const P5: t.PartialT<{
     }>;
 }>;
 export declare const D5: t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.NumberT>>>>>;
-export declare const U1: t.UnionT<[t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const U1: t.UnionT<[t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
         }>;
     }>;
 }>, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.NumberT>>>>>]>;
-export declare const I1: t.IntersectionT<[t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const I1: t.IntersectionT<[t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
         }>;
     }>;
 }>, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.NumberT>>>>>]>;
-export declare const Tu1: t.TupleT<[t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const Tu1: t.TupleT<[t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
         }>;
     }>;
 }>, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.RecordT<t.StringT, t.NumberT>>>>>]>;
-export declare const RO1: t.ReadonlyT<t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const RO1: t.ReadonlyT<t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
         }>;
     }>;
 }>>;
-export declare const RA1: t.ReadonlyArrayT<t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const RA1: t.ReadonlyArrayT<t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;
@@ -115,18 +115,18 @@ export declare const S5: t.StrictT<{
         }>;
     }>;
 }>;
-export declare const TU1: t.TaggedUnionT<"type", [t.InterfaceT<{
+export declare const TU1: t.TaggedUnionT<"type", [t.TypeT<{
     type: t.LiteralT<true>;
     foo: t.StringT;
-}>, t.InterfaceT<{
+}>, t.TypeT<{
     type: t.LiteralT<false>;
     bar: t.NumberT;
 }>]>;
-export declare const E1: t.ExactT<t.InterfaceT<{
-    e: t.InterfaceT<{
-        d: t.InterfaceT<{
-            c: t.InterfaceT<{
-                b: t.InterfaceT<{
+export declare const E1: t.ExactT<t.TypeT<{
+    e: t.TypeT<{
+        d: t.TypeT<{
+            c: t.TypeT<{
+                b: t.TypeT<{
                     a: t.StringT;
                 }>;
             }>;