+++ CHANGELOG.md
@@ -14,6 +14,39 @@
 **Note**: Gaps between patch versions are faulty/broken releases. **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 1.5.3
+
+- **Experimental**
+  - add intermediary interfaces, closes #165 (@gcanti)
+    - `NullC`
+    - `UndefinedC`
+    - `VoidC`
+    - `AnyC`
+    - `UnknownC`
+    - `NeverC`
+    - `StringC`
+    - `NumberC`
+    - `BooleanC`
+    - `UnknownArrayC`
+    - `UnknownRecordC`
+    - `ObjectC`
+    - `FunctionC`
+    - `RefinementC`
+    - `LiteralC`
+    - `KeyofC`
+    - `ArrayC`
+    - `TypeC`
+    - `PartialC`
+    - `RecordC`
+    - `UnionC`
+    - `IntersectionC`
+    - `TupleC`
+    - `ReadonlyC`
+    - `ReadonlyArrayC`
+    - `StrictC`
+    - `TaggedUnionC`
+    - `ExactC`
+
 # 1.5.2
 
 - **Deprecation**